Conflicts:
	.travis.yml

+++ src/server/game/Warden/WardenWin.cpp
@@ -281,7 +281,7 @@ void WardenWin::RequestData()
             }
             case MODULE_CHECK:
             {
-                uint32 seed = static_cast<uint32>(rand32());
+                uint32 seed = rand32();
                 buff << uint32(seed);
                 HmacHash hmac(4, (uint8*)&seed);
                 hmac.UpdateData(wd->Str);
